+++ CHANGELOG.md
@@ -25,7 +25,7 @@ incremented for features.
 * lang: Enforce that the payer for an init-ed account be marked `mut` ([#1271](https://github.com/project-serum/anchor/pull/1271)).
 * lang: All error-related code is now in the error module ([#1426](https://github.com/project-serum/anchor/pull/1426)).
 * lang: Require doc comments when using AccountInfo or UncheckedAccount types ([#1452](https://github.com/project-serum/anchor/pull/1452)).
-* lang: add [`error!`](https://docs.rs/anchor-lang/latest/anchor_lang/prelude/macro.error.html) and [`bail!`](https://docs.rs/anchor-lang/latest/anchor_lang/prelude/macro.bail.html) macro and `Result` type ([#1462](https://github.com/project-serum/anchor/pull/1462)).
+* lang: add [`error!`](https://docs.rs/anchor-lang/latest/anchor_lang/prelude/macro.error.html) and [`err!`](https://docs.rs/anchor-lang/latest/anchor_lang/prelude/macro.err.html) macro and `Result` type ([#1462](https://github.com/project-serum/anchor/pull/1462)).
 This change will break most programs. Do the following to upgrade:
      * change all `ProgramResult`'s to `Result<()>` 
      * change `#[error]` to `#[error_code]`